........
1951 bug fix
........
add empty constructor for hibernate
typo
........
updates to improve performance of connect 
........
#1656 Allow inactive attribute list item in filters
#1656 Allow inactive attribute tree nodes in filters
........
fix for #988
........
#1656 Allow inactive Data Model values in "Group By" and "Values" query filter
........
#1656 Allow inactive list items and tree nodes in "Group By" filter dialog
........
1951 - entity query error
........
1962 fix
........


git-svn-id: https://subversion.assembla.com/svn/smart-cs/trunk@5471 37136b1e-4cb1-4992-8394-573b48a9bdce

+++ source/java/org.wcs.smart.query/src/org/wcs/smart/query/QueryCleanUpJob.java
@@ -50,8 +50,10 @@ public class QueryCleanUpJob extends Job{
 		File dir = QueryPlugIn.getDefault().getQueryTempDirectory();
 		if (dir.exists() && dir.isDirectory()){
 			File[] toDel = dir.listFiles();
-			for (int i = 0; i < toDel.length; i ++){
-				toDel[i].delete();
+			if (toDel != null){
+				for (int i = 0; i < toDel.length; i ++){
+					toDel[i].delete();
+				}
 			}
 		}
